server/requirements.txt
fastapi
uvicorn
requests
pycryptodome
pydantic
